[apps] Avoid unsigned/signed comparison: storeChecksumAtIndex takes an int as parameter

+++ apps/regression/graph_controller.cpp
@@ -351,7 +351,7 @@ uint32_t GraphController::modelVersion() {
   return m_store->storeChecksum();
 }
 
-uint32_t GraphController::modelVersionAtIndex(size_t i) {
+uint32_t GraphController::modelVersionAtIndex(int i) {
   assert(i < numberOfMemoizedVersions());
   return *(m_store->seriesChecksum() + i);
 }
